requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
bs4==0.0.2
cachetools==6.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
curlify==3.0.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
duckduckgo_search==8.1.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
facebook_business==25.0.0
fastapi==0.120.2
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.3
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.6.0
genai==2.1.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.11
ipykernel==6.29.5
ipython==8.35.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.9.2
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langgraph==1.0.1
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.15
langsmith==0.4.14
libclang==18.1.1
lxml==6.0.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
multidict==6.6.3
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.46.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
openai==0.27.10
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.18
ormsgpack==1.12.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.7
playwright==1.55.0
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pycountry==26.2.16
pycparser==3.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
selenium==4.41.0
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.47.1
streamlit==1.53.1
tabulate==0.9.0
tavily==1.1.0
tavily-python==0.7.13
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
threadpoolctl==3.6.0
tiktoken==0.3.3
tinycss2==1.4.0
tokenizers==0.21.4
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.2
trio==0.33.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.3.2
xgboost==3.0.2
xxhash==3.6.0
yarl==1.20.1
youtube-transcript-api==1.2.3
zstandard==0.23.0